+++ queues/templates/queues/edit_queue.html
@@ -47,5 +47,9 @@
         {% endfor %}
     </ul>
 </div>
+{% if not is_guest %}
 <a href="{% url 'profile' %}">Back to Profile</a>
+{%else%}
+Special sign up for Guest user.
+{% endif %}
 {% endblock content %}
