+++ yt_auth/models.py
@@ -1,5 +1,6 @@
 from django.db import models
-from utils import json_to_dict, DjangoFieldsMixin, ToDictMixin
+from utils import json_to_dict
+from mixins import DjangoFieldsMixin, ToDictMixin
 import google.oauth2.credentials as g_oa2_creds
